feat: Update SummaryCard styles and data
The code changes in this commit update the styles and data in the SummaryCard component. The marginTop style is added to create spacing between the SummaryCard and other components. The flex property of the summaryContainer is changed to 2 to allocate more space for the summary text. The textAlign property is set to 'right' to align the summary text to the right. Additionally, the dummy data for expenses is updated with new dates in the year 2024.