refactor: Add deleted property to Expense model
The code changes in this commit add a new 'deleted' property to the Expense model in the 'expense.js' file. This property is used to track whether an expense has been deleted. The 'deleted' property is added to the constructor and initialized with the value 'false' for all existing expenses in the 'dummy-data.js' file.